+++ Changes.md
@@ -11,6 +11,8 @@ Version 3.6.1.0 - TBD
   8. Put back collapsible block functionality.
   9. Improve login page opacity.
  10. Add 'hvpfontcss' setting to allow use of theme fonts within H5P custom CSS.
+ 11. Add Font Awesome 5.11.2 free icons.  Turn on / off with 'fav' setting.
+     Please see the license file 'Font Awesome 5 Free LICENSE.txt' in the fonts folder.
 
 Version 3.6.0.5
   1. Fix core_text scope issue causing exception.